Fix CS1503 in AIInstance inspector binding loop
The fresh-boot pass fed the inspector list index straight into Bind() with a
spurious (uint) cast, but the public binding API (Bind/SetBoundValue) takes an
int slot, so Unity's compiler rejected the uint -> int conversion. Docs already
define the list as 'list index = binding slot', so pass the int index through.

+++ myFSM-UnityRuntime/Runtime/Unity/AIInstance.cs
@@ -221,7 +221,7 @@ namespace MyFSM.Unity
                 for (int i = 0; i < _slotBindings.Count; i++)
                 {
                     if (_slotBindings[i] != null)
-                        Bind((uint)i, _slotBindings[i]);
+                        Bind(i, _slotBindings[i]);
                 }
                 OnBindingsRequired();
             }
